requirements.txt
streamlit>=1.33
streamlit-ace
mistralai>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0
requests>=2.31
watchdog>=3.0
axiom-axle